Refactor bus timer

+++ domain/src/main/java/in/koreatech/koin/domain/model/bus/BusRemainTime.kt
@@ -3,18 +3,18 @@ package `in`.koreatech.koin.domain.model.bus
 sealed class BusRemainTime {
     data class CityBusRemainTime(
         val busNumber: Int,
-        val remainTimeSeconds: Int
+        val remainTimeSeconds: Long
     ) : BusRemainTime()
 
     data class ShuttleBusRemainTime(
-        val remainTimeSeconds: Int
+        val remainTimeSeconds: Long
     ) : BusRemainTime()
 
     data class ExpressBusRemainTime(
-        val remainTimeSeconds: Int
+        val remainTimeSeconds: Long
     ) : BusRemainTime()
 
     data class CommutingBusRemainTime(
-        val remainTimeSeconds: Int
+        val remainTimeSeconds: Long
     ) : BusRemainTime()
 }
